refactor: 사용하지 않는 파일 삭제 및 PostItem 디렉토리 수정
