+++ front-end/legeno-around-here/src/components/PostItem.js
@@ -9,7 +9,7 @@ import Typography from '@material-ui/core/Typography';
 import FavoriteBorderIcon from '@material-ui/icons/FavoriteBorder';
 import FavoriteIcon from '@material-ui/icons/Favorite';
 import CommentIcon from '@material-ui/icons/Comment';
-import { convertDateFormat } from '../../util/TimeUtils';
+import { convertDateFormat } from '../util/TimeUtils';
 
 const useStyles = makeStyles(() => ({
   grow: {
@@ -39,37 +39,23 @@ const useStyles = makeStyles(() => ({
 const PostItem = ({ post }) => {
   const classes = useStyles();
 
-  const {
-    area,
-    commentsCount,
-    createdAt,
-    creator,
-    images,
-    id,
-    zzang,
-    sector,
-    writing,
-  } = post;
+  const { area, commentsCount, createdAt, creator, images, id, zzang, sector, writing } = post;
 
   return (
-    <Card
-      className={classes.root}
-      data-id={id}
-      onClick={() => (document.location.href = `/posts/${id}`)}
-    >
+    <Card className={classes.root} data-id={id} onClick={() => (document.location.href = `/posts/${id}`)}>
       <div className={classes.details}>
         <CardContent className={classes.content}>
-          <Typography component="h6" variant="h5">
+          <Typography component='h6' variant='h5'>
             {sector.name} 부문
           </Typography>
-          <Typography variant="subtitle1" color="textSecondary">
+          <Typography variant='subtitle1' color='textSecondary'>
             {convertDateFormat(createdAt)}
           </Typography>
 
           <Typography
-            variant="body2"
-            color="textSecondary"
-            component="p"
+            variant='body2'
+            color='textSecondary'
+            component='p'
             style={{
               overflow: 'hidden',
               textOverflow: 'ellipsis',
@@ -78,20 +64,16 @@ const PostItem = ({ post }) => {
           >
             {writing}
           </Typography>
-          <Typography variant="subtitle1" color="textSecondary">
+          <Typography variant='subtitle1' color='textSecondary'>
             작성자 : {creator.nickname}
           </Typography>
-          <Typography variant="body2" color="textSecondary" component="p">
+          <Typography variant='body2' color='textSecondary' component='p'>
             작성 지역 : {area.fullName}
           </Typography>
         </CardContent>
         <CardActions disableSpacing>
           <IconButton>
-            {zzang.activated === true ? (
-              <FavoriteIcon />
-            ) : (
-              <FavoriteBorderIcon />
-            )}
+            {zzang.activated === true ? <FavoriteIcon /> : <FavoriteBorderIcon />}
             {zzang.count}
           </IconButton>
           <IconButton>
@@ -101,14 +83,8 @@ const PostItem = ({ post }) => {
         </CardActions>
       </div>
       {images.length > 0 && (
-        <CardMedia
-          className={classes.cover}
-          image={images[0].url}
-          title="Live from space album cover"
-        >
-          {images.length > 1 && (
-            <div className={classes.photoText}>+{images.length - 1}</div>
-          )}
+        <CardMedia className={classes.cover} image={images[0].url} title='Live from space album cover'>
+          {images.length > 1 && <div className={classes.photoText}>+{images.length - 1}</div>}
         </CardMedia>
       )}
     </Card>
