--- front-end/legeno-around-here/src/components/join/Title.js
@@ -1,17 +1,0 @@
-import styled from 'styled-components';
-
-const Title = styled.div`
-  width: 90%;
-  height: 36px;
-  font-family: NotoSansCJKkr;
-  font-size: 24px;
-  font-weight: bold;
-  font-stretch: normal;
-  font-style: normal;
-  line-height: normal;
-  letter-spacing: normal;
-  color: #222222;
-  margin: 15px auto;
-`;
-
-export default Title;
