--- front-end/legeno-around-here/src/components/login/Label.js
@@ -1,18 +1,0 @@
-import styled from "styled-components";
-
-const Label = styled.div`
-  width: 90%;
-  height: 18px;
-  text-align: left;
-  font-family: NotoSansCJKkr;
-  font-size: 12px;
-  font-weight: 500;
-  font-stretch: normal;
-  font-style: normal;
-  line-height: normal;
-  letter-spacing: normal;
-  color: #3f3f3f;
-  font-weight: bold;
-`;
-
-export default Label;
