--- front-end/legeno-around-here/src/components/postdetail/Image.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const Image = styled.img`
-  max-width: 100%;
-  height: auto;
-  box-shadow: 0 0 5px #dddddd;
-  margin-top: 10px;
-  margin-bottom: 15px;
-`;
-
-export default Image;
