--- front-end/legeno-around-here/src/components/posting/ImageInput.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-const ImageInput = styled.input`
-  background-color: #eeeeee;
-  width: 95%;
-  height: 100px;
-  margin: 0 auto 5px auto;
-  text-align: left;
-  padding: 5px;
-`;
-
-export default ImageInput;
